requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
boto3==1.18.23
botocore==1.21.23
click==8.0.1
colorama==0.4.4
decouple==0.0.7
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.11.1
filelock==3.0.12
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
Pillow==8.3.1
platformdirs==2.2.0
psycopg2==2.9.1
pygame==2.0.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.7.2
Werkzeug==2.0.1
whitenoise==5.3.0
